# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: VaultForge API
  version: 0.9.0
  description: |
    HTTP contract for the current VaultForge portfolio application.
    Vault item payloads are encrypted in the browser before they are sent to the API.
servers:
  - url: http://localhost:8080
    description: Local development API
...
paths:
...
  /v1/vaults:
    get:
      operationId: listVaults
      tags: [Vaults]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Owner-scoped vault list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultListResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'

    post:
      operationId: createVault
      tags: [Vaults]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VaultNameRequest'
      responses:
        '201':
          description: Vault created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitedResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
...
  /v1/vaults/{vaultID}/items:
    parameters:
      - $ref: '#/components/parameters/VaultID'

    get:
      operationId: listItems
      tags: [Items]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ItemState'
        - $ref: '#/components/parameters/ItemLimit'
        - $ref: '#/components/parameters/ItemCursor'
      responses:
        '200':
          description: Keyset-paginated owner-scoped item list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemListResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'

    post:
      operationId: createItem
      tags: [Items]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemWriteRequest'
      responses:
        '201':
          $ref: '#/components/responses/ItemCreated'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitedResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
  /v1/vaults/{vaultID}/items/{itemID}:
    parameters:
      - $ref: '#/components/parameters/VaultID'
      - $ref: '#/components/parameters/ItemID'

    get:
      operationId: getItem
      tags: [Items]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ItemState'
      responses:
        '200':
          $ref: '#/components/responses/ItemOK'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'

    put:
      operationId: updateItem
      tags: [Items]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemWriteRequest'
      responses:
        '200':
          $ref: '#/components/responses/ItemOK'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '412':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitedResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'

    delete:
      operationId: softDeleteItem
      tags: [Items]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          $ref: '#/components/responses/ItemOK'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '412':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitedResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
  /v1/vaults/{vaultID}/items/{itemID}/restore:
    parameters:
      - $ref: '#/components/parameters/VaultID'
      - $ref: '#/components/parameters/ItemID'

    post:
      operationId: restoreItem
      tags: [Items]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          $ref: '#/components/responses/ItemOK'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '412':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitedResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    VaultID:
      name: vaultID
      in: path
      required: true
      schema:
        type: string
        maxLength: 256
    ItemID:
      name: itemID
      in: path
      required: true
      schema:
        type: string
        maxLength: 256
    ItemState:
      name: state
      in: query
      required: false
      schema:
        type: string
        enum: [active, deleted]
    ItemLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
    ItemCursor:
      name: after
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 1024
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 255
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong quoted positive item version such as `"3"`.
      schema:
        type: string
        minLength: 3
        maxLength: 32
        pattern: '^"[1-9][0-9]*"$'
  responses:
...
    ErrorResponse:
      description: Safe public error envelope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RateLimitedResponse:
      description: Request limit exceeded.
      headers:
        Retry-After:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    ItemCreated:
      description: Item created or replayed from durable idempotency state.
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ItemResponse'
    ItemOK:
      description: Current item representation.
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ItemResponse'
  headers:
    ETag:
      description: Strong quoted item version.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
            - message
          properties:
            code:
              type: string
              minLength: 1
            message:
              type: string
              minLength: 1
            request_id:
              type: string
              minLength: 1
...
    VaultNameRequest:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
...
    Vault:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        name:
          type: string
        cryptoVersion:
          type: integer
          format: int32
          nullable: true
        kdfVersion:
          type: integer
          format: int32
          nullable: true
        salt:
          type: string
          format: byte
        wrappedKey:
          type: string
          format: byte
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    VaultResponse:
      type: object
      additionalProperties: false
      required:
        - vault
      properties:
        vault:
          $ref: '#/components/schemas/Vault'
    VaultListResponse:
      type: object
      additionalProperties: false
      required:
        - vaults
      properties:
        vaults:
          type: array
          items:
            $ref: '#/components/schemas/Vault'
    ItemType:
      type: string
      enum:
        - login
        - api_key
        - environment_variable
        - database_connection
        - secure_note
    ItemEncryptedPayload:
      type: object
      additionalProperties: false
      required:
        - version
        - algorithm
        - blob
      properties:
        version:
          type: integer
          format: int32
          enum:
            - 1
        algorithm:
          type: string
          enum:
            - AES-256-GCM
        blob:
          type: string
          format: byte
          minLength: 1
          maxLength: 90000
    ItemWriteRequest:
      type: object
      additionalProperties: false
      required:
        - type
        - encryptedPayload
      properties:
        type:
          $ref: '#/components/schemas/ItemType'
        encryptedPayload:
          $ref: '#/components/schemas/ItemEncryptedPayload'
    Item:
      type: object
      additionalProperties: false
      required:
        - id
        - type
        - encryptedPayload
        - version
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/ItemType'
        encryptedPayload:
          $ref: '#/components/schemas/ItemEncryptedPayload'
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
    ItemResponse:
      type: object
      additionalProperties: false
      required:
        - item
      properties:
        item:
          $ref: '#/components/schemas/Item'
    ItemListResponse:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        nextCursor:
          type: string
          minLength: 1
          maxLength: 1024